projects/databricks-genie-multiagent/azure-teams-deployment/requirements.txt
requests
python-dotenv
pandas
botbuilder
aiohttp
botbuilder-core
botbuilder-ai
botframework-connector
botbuilder-schema
botbuilder-dialogs
botbuilder-integration-aiohttp
databricks-sdk
azure-identity